requirements.txt
mutagen==1.45.1
PyYAML==6.0
